+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 
 - feat:support spring cloud 2024.
 - [fix: add gateway context config example.](https://github.com/Tencent/spring-cloud-tencent/pull/1584)
+- [feat:support config empty protection.](https://github.com/Tencent/spring-cloud-tencent/pull/1587)